+++ dart/lib/src/noop_hub.dart
@@ -4,6 +4,7 @@ import 'package:meta/meta.dart';
 
 import 'hint.dart';
 import 'hub.dart';
+import 'profiling.dart';
 import 'protocol.dart';
 import 'scope.dart';
 import 'sentry_client.dart';
@@ -119,6 +120,14 @@ class NoOpHub implements Hub {
 
   @override
   void setSpanContext(throwable, ISentrySpan span, String transaction) {}
+
+  @internal
+  @override
+  set profilerFactory(SentryProfilerFactory? value) {}
+
+  @internal
+  @override
+  SentryProfilerFactory? get profilerFactory => null;
 
   @override
   Scope get scope => Scope(_options);