+++ dart/lib/src/protocol/sentry_app.dart
@@ -18,6 +18,7 @@ class SentryApp {
     this.deviceAppHash,
     this.appMemory,
     this.inForeground,
+    this.viewNames,
   });
 
   /// Human readable application name, as it appears on the platform.
@@ -48,20 +49,27 @@ class SentryApp {
   /// An app is in foreground when it's visible to the user.
   final bool? inForeground;
 
+  /// The names of the currently visible views.
+  final List<String>? viewNames;
+
   /// Deserializes a [SentryApp] from JSON [Map].
-  factory SentryApp.fromJson(Map<String, dynamic> data) => SentryApp(
-        name: data['app_name'],
-        version: data['app_version'],
-        identifier: data['app_identifier'],
-        build: data['app_build'],
-        buildType: data['build_type'],
-        startTime: data['app_start_time'] != null
-            ? DateTime.tryParse(data['app_start_time'])
-            : null,
-        deviceAppHash: data['device_app_hash'],
-        appMemory: data['app_memory'],
-        inForeground: data['in_foreground'],
-      );
+  factory SentryApp.fromJson(Map<String, dynamic> data) {
+    final viewNamesJson = data['view_names'] as List<dynamic>?;
+    return SentryApp(
+      name: data['app_name'],
+      version: data['app_version'],
+      identifier: data['app_identifier'],
+      build: data['app_build'],
+      buildType: data['build_type'],
+      startTime: data['app_start_time'] != null
+          ? DateTime.tryParse(data['app_start_time'])
+          : null,
+      deviceAppHash: data['device_app_hash'],
+      appMemory: data['app_memory'],
+      inForeground: data['in_foreground'],
+      viewNames: viewNamesJson?.map((e) => e as String).toList(),
+    );
+  }
 
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
@@ -71,10 +79,11 @@ class SentryApp {
       if (identifier != null) 'app_identifier': identifier!,
       if (build != null) 'app_build': build!,
       if (buildType != null) 'build_type': buildType!,
+      if (startTime != null) 'app_start_time': startTime!.toIso8601String(),
       if (deviceAppHash != null) 'device_app_hash': deviceAppHash!,
       if (appMemory != null) 'app_memory': appMemory!,
-      if (startTime != null) 'app_start_time': startTime!.toIso8601String(),
       if (inForeground != null) 'in_foreground': inForeground!,
+      if (viewNames != null && viewNames!.isNotEmpty) 'view_names': viewNames!,
     };
   }
 
@@ -88,6 +97,7 @@ class SentryApp {
         deviceAppHash: deviceAppHash,
         appMemory: appMemory,
         inForeground: inForeground,
+        viewNames: viewNames,
       );
 
   SentryApp copyWith({
@@ -100,6 +110,7 @@ class SentryApp {
     String? deviceAppHash,
     int? appMemory,
     bool? inForeground,
+    List<String>? viewNames,
   }) =>
       SentryApp(
         name: name ?? this.name,
@@ -111,5 +122,6 @@ class SentryApp {
         deviceAppHash: deviceAppHash ?? this.deviceAppHash,
         appMemory: appMemory ?? this.appMemory,
         inForeground: inForeground ?? this.inForeground,
+        viewNames: viewNames ?? this.viewNames,
       );
 }