+++ flutter/lib/src/native/sentry_native.dart
@@ -3,28 +3,17 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 
 import '../../sentry_flutter.dart';
-import 'sentry_native_channel.dart';
+import 'sentry_native_binding.dart';
 
-/// [SentryNative] holds state that it fetches from to the native SDKs. Always
-/// use the shared instance with [SentryNative()].
+/// [SentryNative] holds state that it fetches from to the native SDKs.
+/// It forwards to platform-specific implementations of [SentryNativeBinding].
+/// Any errors are logged and ignored.
 @internal
 class SentryNative {
-  SentryNative._();
+  final SentryOptions _options;
+  final SentryNativeBinding _binding;
 
-  static final SentryNative _instance = SentryNative._();
-
-  SentryNativeChannel? _nativeChannel;
-
-  factory SentryNative() {
-    return _instance;
-  }
-
-  SentryNativeChannel? get nativeChannel => _instance._nativeChannel;
-
-  /// Provide [nativeChannel] for native communication.
-  set nativeChannel(SentryNativeChannel? nativeChannel) {
-    _instance._nativeChannel = nativeChannel;
-  }
+  SentryNative(this._options, this._binding);
 
   // AppStart
 
@@ -41,61 +30,98 @@ class SentryNative {
   /// Fetch [NativeAppStart] from native channels. Can only be called once.
   Future<NativeAppStart?> fetchNativeAppStart() async {
     _didFetchAppStart = true;
-    return await _nativeChannel?.fetchNativeAppStart();
+    return _invoke("fetchNativeAppStart", _binding.fetchNativeAppStart);
   }
 
   // NativeFrames
 
-  Future<void> beginNativeFramesCollection() async {
-    await _nativeChannel?.beginNativeFrames();
-  }
+  Future<void> beginNativeFramesCollection() =>
+      _invoke("beginNativeFrames", _binding.beginNativeFrames);
 
-  Future<NativeFrames?> endNativeFramesCollection(SentryId traceId) async {
-    return await _nativeChannel?.endNativeFrames(traceId);
-  }
+  Future<NativeFrames?> endNativeFramesCollection(SentryId traceId) =>
+      _invoke("endNativeFrames", () => _binding.endNativeFrames(traceId));
 
   // Scope
 
-  Future<void> setContexts(String key, dynamic value) async {
-    return await _nativeChannel?.setContexts(key, value);
-  }
+  Future<void> setContexts(String key, dynamic value) =>
+      _invoke("setContexts", () => _binding.setContexts(key, value));
 
-  Future<void> removeContexts(String key) async {
-    return await _nativeChannel?.removeContexts(key);
-  }
+  Future<void> removeContexts(String key) =>
+      _invoke("removeContexts", () => _binding.removeContexts(key));
 
-  Future<void> setUser(SentryUser? sentryUser) async {
-    return await _nativeChannel?.setUser(sentryUser);
-  }
+  Future<void> setUser(SentryUser? sentryUser) =>
+      _invoke("setUser", () => _binding.setUser(sentryUser));
 
-  Future<void> addBreadcrumb(Breadcrumb breadcrumb) async {
-    return await _nativeChannel?.addBreadcrumb(breadcrumb);
-  }
+  Future<void> addBreadcrumb(Breadcrumb breadcrumb) =>
+      _invoke("addBreadcrumb", () => _binding.addBreadcrumb(breadcrumb));
 
-  Future<void> clearBreadcrumbs() async {
-    return await _nativeChannel?.clearBreadcrumbs();
-  }
+  Future<void> clearBreadcrumbs() =>
+      _invoke("clearBreadcrumbs", _binding.clearBreadcrumbs);
 
-  Future<void> setExtra(String key, dynamic value) async {
-    return await _nativeChannel?.setExtra(key, value);
-  }
+  Future<void> setExtra(String key, dynamic value) =>
+      _invoke("setExtra", () => _binding.setExtra(key, value));
 
-  Future<void> removeExtra(String key) async {
-    return await _nativeChannel?.removeExtra(key);
-  }
+  Future<void> removeExtra(String key) =>
+      _invoke("removeExtra", () => _binding.removeExtra(key));
 
-  Future<void> setTag(String key, String value) async {
-    return await _nativeChannel?.setTag(key, value);
-  }
+  Future<void> setTag(String key, String value) =>
+      _invoke("setTag", () => _binding.setTag(key, value));
 
-  Future<void> removeTag(String key) async {
-    return await _nativeChannel?.removeTag(key);
-  }
+  Future<void> removeTag(String key) =>
+      _invoke("removeTag", () => _binding.removeTag(key));
+
+  int? startProfiler(SentryId traceId) =>
+      _invokeSync("startProfiler", () => _binding.startProfiler(traceId));
+
+  Future<void> discardProfiler(SentryId traceId) =>
+      _invoke("discardProfiler", () => _binding.discardProfiler(traceId));
+
+  Future<Map<String, dynamic>?> collectProfile(
+          SentryId traceId, int startTimeNs, int endTimeNs) =>
+      _invoke("collectProfile",
+          () => _binding.collectProfile(traceId, startTimeNs, endTimeNs));
 
   /// Reset state
   void reset() {
     appStartEnd = null;
     _didFetchAppStart = false;
+  }
+
+  // Helpers
+  Future<T?> _invoke<T>(
+      String nativeMethodName, Future<T?> Function() fn) async {
+    try {
+      return await fn();
+    } catch (error, stackTrace) {
+      _logError(nativeMethodName, error, stackTrace);
+      // ignore: invalid_use_of_internal_member
+      if (_options.automatedTestMode) {
+        rethrow;
+      }
+      return null;
+    }
+  }
+
+  T? _invokeSync<T>(String nativeMethodName, T? Function() fn) {
+    try {
+      return fn();
+    } catch (error, stackTrace) {
+      _logError(nativeMethodName, error, stackTrace);
+      // ignore: invalid_use_of_internal_member
+      if (_options.automatedTestMode) {
+        rethrow;
+      }
+      return null;
+    }
+  }
+
+  void _logError(String nativeMethodName, Object error, StackTrace stackTrace) {
+    _options.logger(
+      SentryLevel.error,
+      'Native call `$nativeMethodName` failed',
+      exception: error,
+      stackTrace: stackTrace,
+    );
   }
 }