+++ flutter/lib/src/sentry_flutter.dart
@@ -10,10 +10,11 @@ import 'event_processor/android_platform_exception_event_processor.dart';
 import 'event_processor/flutter_exception_event_processor.dart';
 import 'event_processor/platform_exception_event_processor.dart';
 import 'integrations/screenshot_integration.dart';
+import 'native/factory.dart';
 import 'native/native_scope_observer.dart';
+import 'profiling.dart';
 import 'renderer/renderer.dart';
 import 'native/sentry_native.dart';
-import 'native/sentry_native_channel.dart';
 
 import 'integrations/integrations.dart';
 import 'event_processor/flutter_enricher_event_processor.dart';
@@ -48,9 +49,8 @@ mixin SentryFlutter {
     }
 
     if (flutterOptions.platformChecker.hasNativeIntegration) {
-      // Set a default native channel to the singleton SentryNative instance.
-      SentryNative().nativeChannel =
-          SentryNativeChannel(channel, flutterOptions);
+      final binding = createBinding(flutterOptions.platformChecker, channel);
+      _native = SentryNative(flutterOptions, binding);
     }
 
     final platformDispatcher = PlatformDispatcher.instance;
@@ -81,9 +81,7 @@ mixin SentryFlutter {
     await _initDefaultValues(flutterOptions, channel);
 
     await Sentry.init(
-      (options) async {
-        await optionsConfiguration(options as SentryFlutterOptions);
-      },
+      (options) => optionsConfiguration(options as SentryFlutterOptions),
       appRunner: appRunner,
       // ignore: invalid_use_of_internal_member
       options: flutterOptions,
@@ -92,6 +90,11 @@ mixin SentryFlutter {
       // ignore: invalid_use_of_internal_member
       runZonedGuardedOnError: runZonedGuardedOnError,
     );
+
+    if (_native != null) {
+      // ignore: invalid_use_of_internal_member
+      SentryNativeProfilerFactory.attachTo(Sentry.currentHub, _native!);
+    }
   }
 
   static Future<void> _initDefaultValues(
@@ -101,9 +104,9 @@ mixin SentryFlutter {
     options.addEventProcessor(FlutterExceptionEventProcessor());
 
     // Not all platforms have a native integration.
-    if (options.platformChecker.hasNativeIntegration) {
+    if (_native != null) {
       options.transport = FileSystemTransport(channel, options);
-      options.addScopeObserver(NativeScopeObserver(SentryNative()));
+      options.addScopeObserver(NativeScopeObserver(_native!));
     }
 
     var flutterEventProcessor = FlutterEnricherEventProcessor(options);
@@ -179,9 +182,9 @@ mixin SentryFlutter {
     // in errors.
     integrations.add(LoadReleaseIntegration());
 
-    if (platformChecker.hasNativeIntegration) {
+    if (_native != null) {
       integrations.add(NativeAppStartIntegration(
-        SentryNative(),
+        _native!,
         () {
           try {
             /// Flutter >= 2.12 throws if SchedulerBinding.instance isn't initialized.
@@ -207,7 +210,7 @@ mixin SentryFlutter {
   /// Manually set when your app finished startup. Make sure to set
   /// [SentryFlutterOptions.autoAppStart] to false on init.
   static void setAppStartEnd(DateTime appStartEnd) {
-    SentryNative().appStartEnd = appStartEnd;
+    _native?.appStartEnd = appStartEnd;
   }
 
   static void _setSdk(SentryFlutterOptions options) {
@@ -221,4 +224,10 @@ mixin SentryFlutter {
     sdk.addPackage('pub:sentry_flutter', sdkVersion);
     options.sdk = sdk;
   }
+
+  @internal
+  static SentryNative? get native => _native;
+  @internal
+  static set native(SentryNative? value) => _native = value;
+  static SentryNative? _native;
 }