+++ sqflite/test/sentry_database_test.dart
@@ -26,6 +26,7 @@ void main() {
       fixture = Fixture();
 
       when(fixture.hub.options).thenReturn(fixture.options);
+      when(fixture.hub.scope).thenReturn(fixture.scope);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
 
       // using ffi for testing on vm
@@ -74,6 +75,17 @@ void main() {
       );
     });
 
+    test('creates close breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.close();
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.message, 'Close DB: $inMemoryDatabasePath');
+      expect(breadcrumb.category, SentryDatabase.dbOp);
+      expect(breadcrumb.type, 'query');
+    });
+
     test('creates transaction span', () async {
       final db = await fixture.getSut();
 
@@ -91,6 +103,27 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabase,
       );
+
+      await db.close();
+    });
+
+    test('creates transaction breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.transaction((txn) async {
+        expect(txn is SentrySqfliteTransaction, true);
+      });
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.message, 'Transaction DB: $inMemoryDatabasePath');
+      expect(breadcrumb.category, 'db.sql.transaction');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -177,6 +210,7 @@ void main() {
       fixture = Fixture();
 
       when(fixture.hub.options).thenReturn(fixture.options);
+      when(fixture.hub.scope).thenReturn(fixture.scope);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
 
       when(fixture.database.path).thenReturn('/path/db');
@@ -201,6 +235,19 @@ void main() {
       );
     });
 
+    test('close sets breadcrumb to internal error', () async {
+      when(fixture.database.close()).thenThrow(fixture.exception);
+
+      final db = await fixture.getSut(database: fixture.database);
+
+      await expectLater(() async => await db.close(), throwsException);
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
     test('transaction sets span to internal error', () async {
       // ignore: inference_failure_on_function_invocation
       when(fixture.database.transaction(any)).thenThrow(fixture.exception);
@@ -222,6 +269,23 @@ void main() {
         SentryTraceOrigins.autoDbSqfliteDatabase,
       );
     });
+
+    test('transaction sets breadcrumb to internal error', () async {
+      // ignore: inference_failure_on_function_invocation
+      when(fixture.database.transaction(any)).thenThrow(fixture.exception);
+
+      final db = await fixture.getSut(database: fixture.database);
+
+      await expectLater(
+        () async => await db.transaction((txn) async {}),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
   });
 
   group('$SentryDatabaseExecutor success', () {
@@ -231,6 +295,7 @@ void main() {
       fixture = Fixture();
 
       when(fixture.hub.options).thenReturn(fixture.options);
+      when(fixture.hub.scope).thenReturn(fixture.scope);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
 
       // using ffi for testing on vm
@@ -259,6 +324,25 @@ void main() {
       await db.close();
     });
 
+    test('creates delete breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.delete('Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'DELETE FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     test('creates execute span', () async {
       final db = await fixture.getSut();
 
@@ -276,6 +360,25 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
+
+      await db.close();
+    });
+
+    test('creates execute breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.execute('DELETE FROM Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'DELETE FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -304,6 +407,28 @@ void main() {
       await db.close();
     });
 
+    test('creates insert breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.insert('Product', <String, Object?>{'title': 'Product 1'});
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(
+        breadcrumb.message,
+        'INSERT INTO Product (title) VALUES (?)',
+      );
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     test('creates query span', () async {
       final db = await fixture.getSut();
 
@@ -321,6 +446,26 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
+
+      await db.close();
+    });
+
+    test('creates query breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.query('Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.query');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.message, 'SELECT * FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -346,6 +491,26 @@ void main() {
       await db.close();
     });
 
+    test('creates query cursor breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.queryCursor('Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.query');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.message, 'SELECT * FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     test('creates raw delete span', () async {
       final db = await fixture.getSut();
 
@@ -363,6 +528,25 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
+
+      await db.close();
+    });
+
+    test('creates raw delete breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.rawDelete('DELETE FROM Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'DELETE FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -392,6 +576,26 @@ void main() {
       await db.close();
     });
 
+    test('creates raw insert breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db
+          .rawInsert('INSERT INTO Product (title) VALUES (?)', ['Product 1']);
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'INSERT INTO Product (title) VALUES (?)');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     test('creates raw query span', () async {
       final db = await fixture.getSut();
 
@@ -409,6 +613,26 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
+
+      await db.close();
+    });
+
+    test('creates raw query breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.rawQuery('SELECT * FROM Product');
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.query');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.message, 'SELECT * FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -434,6 +658,26 @@ void main() {
       await db.close();
     });
 
+    test('creates raw query cursor breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.rawQueryCursor('SELECT * FROM Product', []);
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.query');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.message, 'SELECT * FROM Product');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     test('creates raw update span', () async {
       final db = await fixture.getSut();
 
@@ -451,6 +695,25 @@ void main() {
         // ignore: invalid_use_of_internal_member
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
+
+      await db.close();
+    });
+
+    test('creates raw update breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.rawUpdate('UPDATE Product SET title = ?', ['Product 1']);
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'UPDATE Product SET title = ?');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
 
       await db.close();
     });
@@ -476,6 +739,25 @@ void main() {
       await db.close();
     });
 
+    test('creates update breadcrumb', () async {
+      final db = await fixture.getSut();
+
+      await db.update('Product', <String, Object?>{'title': 'Product 1'});
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.category, 'db.sql.execute');
+      expect(breadcrumb.message, 'UPDATE Product SET title = ?');
+      expect(breadcrumb.data?['status'], 'ok');
+      expect(
+        breadcrumb.data?[SentryDatabase.dbSystemKey],
+        SentryDatabase.dbSystem,
+      );
+      expect(breadcrumb.data?[SentryDatabase.dbNameKey], inMemoryDatabasePath);
+      expect(breadcrumb.type, 'query');
+
+      await db.close();
+    });
+
     tearDown(() {
       databaseFactory = sqfliteDatabaseFactoryDefault;
     });
@@ -488,6 +770,7 @@ void main() {
       fixture = Fixture();
 
       when(fixture.hub.options).thenReturn(fixture.options);
+      when(fixture.hub.scope).thenReturn(fixture.scope);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
     });
 
@@ -724,6 +1007,185 @@ void main() {
         SentryTraceOrigins.autoDbSqfliteDatabaseExecutor,
       );
     });
+
+    test('delete sets breadcrumb to internal error', () async {
+      when(fixture.executor.delete(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.delete('Product'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('execute sets breadcrumb to internal error', () async {
+      when(fixture.executor.execute(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.execute('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('insert sets breadcrumb to internal error', () async {
+      when(fixture.executor.insert(any, any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor
+            .insert('Product', <String, Object?>{'title': 'Product 1'}),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('query sets breadcrumb to internal error', () async {
+      when(fixture.executor.query(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.query('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('query cursor sets breadcrumb to internal error', () async {
+      when(fixture.executor.queryCursor(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.queryCursor('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('raw delete sets breadcrumb to internal error', () async {
+      when(fixture.executor.rawDelete(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.rawDelete('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('raw insert sets breadcrumb to internal error', () async {
+      when(fixture.executor.rawInsert(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.rawInsert('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('raw query sets breadcrumb to internal error', () async {
+      when(fixture.executor.rawQuery(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.rawQuery('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('raw query cursor sets breadcrumb to internal error', () async {
+      when(fixture.executor.rawQueryCursor(any, any))
+          .thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.rawQueryCursor('sql', []),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('raw update sets breadcrumb to internal error', () async {
+      when(fixture.executor.rawUpdate(any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor.rawUpdate('sql'),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
+
+    test('update sets breadcrumb to internal error', () async {
+      when(fixture.executor.update(any, any)).thenThrow(fixture.exception);
+
+      final executor = fixture.getExecutorSut();
+
+      await expectLater(
+        () async => await executor
+            .update('Product', <String, Object?>{'title': 'Product 1'}),
+        throwsException,
+      );
+
+      final breadcrumb = fixture.hub.scope.breadcrumbs.first;
+      expect(breadcrumb.data?['status'], 'internal_error');
+      expect(breadcrumb.type, 'query');
+      expect(breadcrumb.level, SentryLevel.warning);
+    });
   });
 }
 
@@ -735,6 +1197,7 @@ class Fixture {
   final database = MockDatabase();
   final exception = Exception('error');
   final executor = MockDatabaseExecutor();
+  late final scope = Scope(options);
 
   Future<SentryDatabase> getSut({
     double? tracesSampleRate = 1.0,